requirements.txt
absl-py
astunparse
cachetools
certifi
charset-normalizer
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
keras
Keras-Preprocessing
libclang
Markdown
numpy
oauthlib
opt-einsum
packaging
pillow
protobuf
pyasn1
pyasn1-modules
requests
requests-oauthlib
rsa
six
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
termcolor
urllib3
Werkzeug
wrapt
